Rama irene

+++ main/src/main/java/com/main/comunicacion/privadas/DTOs/CiudadPrivateDTO.java
@@ -6,12 +6,12 @@ import lombok.Data;
 
 @Data
 public class CiudadPrivateDTO {
-    private Integer id;
+    private long id;
     private String nombre;
     private String latitud;
     private String longitud;
     
-    public CiudadPrivateDTO(Integer id, String nombre, String latitud, String longitud) {
+    public CiudadPrivateDTO(long id, String nombre, String latitud, String longitud) {
         this.id = id;
         this.nombre = nombre;
         this.latitud = latitud;
